+++ shared/store/src/index.ts
@@ -1,9 +1,8 @@
-export * from './lib/App';
 export { PageOne } from './lib/components/navigation-testing/PageOne';
 export { PageThree } from './lib/components/navigation-testing/PageThree';
 export { PageTwo } from './lib/components/navigation-testing/PageTwo';
 export { Counter } from './lib/components/store-testing/Counter';
-export * from './lib/rufferal-design/system';
+export * from './lib/rufferal-design/components';
 export * from './store';
 
 // import { GoogleSignin } from '@react-native-google-signin/google-signin';
