+++ shared/store/src/lib/components/navigation-testing/PageThree.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Text } from 'react-native';
-import { AccessTemplate } from '../../rufferal-design/system/templates/AccessTemplate';
+import { AccessTemplate } from '../../rufferal-design/components/templates/AccessTemplate';
 
 export const PageThree = (): React.ReactElement => {
   return (
